Move Chunk struct to file of its own
